Extract RuntimeCore from Scheduler, unify audio/cadence API
Refactor the runtime to extract a RuntimeCore that owns the Graph,
OperatorRegistry, CompiledGraph, CadenceBridge, and FrameExecutor.
Scheduler and AudioEngine now operate on RuntimeCore, simplifying
their APIs (e.g. audio->build(sched.core())). Rename "domain" to
"cadence" throughout (test_audio_domain_sequencer → test_audio_cadence_sequencer,
test_cross_domain_spread → test_cross_cadence_spread).

+++ src/runtime/control_server.cpp
@@ -598,7 +598,7 @@ static std::string handle_introspect_nodes(Graph& graph, Scheduler& scheduler) {
         node["cadence_capability"] = (ns.cadence_capability == VIVID_CADENCE_AUDIO_CAPABLE) ? "audio_capable" : "frame_only";
         {
             const auto* ndef = graph.find_node(ns.node_id);
-            node["cadence_override"] = ndef ? static_cast<int>(ndef->cadence_override) : 0;
+            node["cadence_override"] = ndef ? static_cast<int>(static_cast<uint8_t>(ndef->cadence_override)) : 0;
         }
         node["incoming_wires"] = static_cast<int64_t>(incoming_wires[ns.node_id]);
         node["outgoing_wires"] = static_cast<int64_t>(outgoing_wires[ns.node_id]);
